moved serde code out of variant.rs
