+++ parquet-variant/src/from_json.rs
@@ -1,7 +1,9 @@
+pub use crate::variant::{VariantDecimal4, VariantDecimal8};
 use crate::variant_buffer_manager::VariantBufferManager;
 use crate::{AppendVariantHelper, ListBuilder, ObjectBuilder, Variant, VariantBuilder};
 use arrow_schema::ArrowError;
-use serde_json::{Map, Value};
+use rust_decimal::prelude::*;
+use serde_json::{Map, Number, Value};
 
 /// Eventually, internal writes should also be performed using VariantBufferManager instead of
 /// ValueBuffer and MetadataBuffer so the caller has control of the memory.
@@ -35,11 +37,59 @@ fn build_json(json: &Value, builder: &mut VariantBuilder) -> Result<(), ArrowErr
     Ok(())
 }
 
+fn variant_from_number<'a, 'b>(n: &Number) -> Result<Variant<'a, 'b>, ArrowError> {
+    if let Some(i) = n.as_i64() {
+        // Find minimum Integer width to fit
+        if i as i8 as i64 == i {
+            Ok((i as i8).into())
+        } else if i as i16 as i64 == i {
+            Ok((i as i16).into())
+        } else if i as i32 as i64 == i {
+            Ok((i as i32).into())
+        } else {
+            Ok(i.into())
+        }
+    } else {
+        // Try decimal
+        // TODO: Replace with custom decimal parsing as the rust_decimal library only supports
+        // a max unscaled value of 2^96.
+        match Decimal::from_str_exact(n.as_str()) {
+            Ok(dec) => {
+                let unscaled: i128 = dec.mantissa();
+                let scale = dec.scale() as u8;
+                if unscaled.abs() <= VariantDecimal4::MAX_UNSCALED_VALUE as i128
+                    && scale <= VariantDecimal4::MAX_PRECISION as u8
+                {
+                    (unscaled as i32, scale).try_into()
+                } else if unscaled.abs() <= VariantDecimal8::MAX_UNSCALED_VALUE as i128
+                    && scale <= VariantDecimal8::MAX_PRECISION as u8
+                {
+                    (unscaled as i64, scale).try_into()
+                } else {
+                    (unscaled, scale).try_into()
+                }
+            }
+            Err(_) => {
+                // Try double
+                match n.as_f64() {
+                    Some(f) => return Ok(f.into()),
+                    None => Err(ArrowError::InvalidArgumentError(format!(
+                        "Failed to parse {} as number",
+                        n.as_str()
+                    ))),
+                }?
+            }
+        }
+    }
+}
+
 fn append_json(json: &Value, builder: &mut impl AppendVariantHelper) -> Result<(), ArrowError> {
     match json {
         Value::Null => builder.append_value_helper(Variant::Null),
         Value::Bool(b) => builder.append_value_helper(*b),
-        Value::Number(n) => builder.append_value_helper(Variant::try_from(n)?),
+        Value::Number(n) => {
+            builder.append_value_helper(variant_from_number(n)?);
+        }
         Value::String(s) => builder.append_value_helper(s.as_str()),
         Value::Array(arr) => {
             let mut list_builder = builder.new_list_helper();
